+++ src/src/component/Setting.js
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
 import { NavLink } from 'react-router-dom';
 import { useSpring, animated } from 'react-spring';
-import Profile from './Profile';
-import Email from './Email';
-import Password from './Password';
-import Withdraw from './Withdraw';
+import Profile from './SettingProfile';
+import Email from './SettingEmail';
+import Password from './SettingPassword';
+import Withdraw from './SettingWithdraw';
 
 export default function Setting() {
   const initialIndex = () => {
